# aidis-command/frontend/.openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIDIS Command Backend API",
    "version": "1.0.0",
    "description": "RESTful API for AIDIS Command Backend with MCP Bridge Integration",
    "contact": {
      "name": "AIDIS Development Team",
      "email": "dev@aidis.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3002/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data payload"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation ID for tracking"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "version": {
                "type": "string",
                "example": "1.0.0"
              }
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "BaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ]
      },
...
      "DecisionEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "problem": {
                "type": "string"
              },
              "decision": {
                "type": "string"
              },
              "rationale": {
                "type": "string"
              },
              "alternatives": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "under_review",
                  "superseded",
                  "deprecated"
                ]
              },
              "outcomeStatus": {
                "type": "string",
                "enum": [
                  "unknown",
                  "successful",
                  "failed",
                  "mixed",
                  "too_early"
                ],
                "description": "Lifecycle outcome assessment"
              },
              "outcomeNotes": {
                "type": "string"
              },
              "lessonsLearned": {
                "type": "string"
              },
              "supersededBy": {
                "type": "string"
              },
              "supersededReason": {
                "type": "string"
              },
              "outcome": {
                "type": "string"
              },
              "lessons": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "project_id": {
                "type": "string",
                "format": "uuid"
              },
              "project_name": {
                "type": "string"
              },
              "created_by": {
                "type": "string"
              },
              "updated_by": {
                "type": "string"
              }
            },
            "required": [
              "title",
              "problem",
              "decision",
              "status"
            ]
          }
        ]
      },
      "DecisionSearchResponse": {
        "type": "object",
        "properties": {
          "decisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DecisionEntity"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        },
        "required": [
          "decisions",
          "total"
        ]
      },
      "DecisionStats": {
        "type": "object",
        "properties": {
          "total_decisions": {
            "type": "integer"
          },
          "by_status": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "by_project": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "recent_decisions": {
            "type": "integer"
          },
          "total_projects": {
            "type": "integer"
          }
        }
      },
      "CreateDecisionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "problem": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "decision": {
            "type": "string",
            "minLength": 10,
            "maxLength": 5000
          },
          "rationale": {
            "type": "string",
            "maxLength": 5000
          },
          "alternatives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "under_review",
              "superseded",
              "deprecated"
            ],
            "default": "active"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "problem",
          "decision"
        ]
      },
      "UpdateDecisionRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "problem": {
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "alternatives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "under_review",
              "superseded",
              "deprecated"
            ]
          },
          "outcomeStatus": {
            "type": "string",
            "enum": [
              "unknown",
              "successful",
              "failed",
              "mixed",
              "too_early"
            ]
          },
          "outcomeNotes": {
            "type": "string"
          },
          "lessonsLearned": {
            "type": "string"
          },
          "supersededReason": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/decisions": {
      "get": {
        "summary": "Search technical decisions",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Free-text search term to locate decisions"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "under_review",
                "superseded",
                "deprecated"
              ]
            },
            "description": "Filter decisions by lifecycle status"
          },
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Scope search to a specific project"
          },
          {
            "in": "query",
            "name": "created_by",
            "schema": {
              "type": "string"
            },
            "description": "Filter by author username"
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive lower bound for creation date"
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Inclusive upper bound for creation date"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decisions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DecisionSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Record a new technical decision",
        "tags": [
          "Decisions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Decision recorded successfully"
          }
        }
      }
    },
    "/decisions/stats": {
      "get": {
        "summary": "Retrieve decision statistics",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Optional project scope for statistics"
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics returned",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DecisionStats"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/decisions/{id}": {
      "get": {
        "summary": "Get a decision by ID",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decision retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DecisionEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Decision not found"
          }
        }
      },
      "put": {
        "summary": "Update a decision",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decision updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a decision",
        "tags": [
          "Decisions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decision deleted successfully"
          }
        }
      }
    },
...
  }
}